--- a/app/models/rule_references_rule.rb
+++ b/app/models/rule_references_rule.rb
@@ -7,4 +7,14 @@ class RuleReferencesRule < ApplicationRecord
 
   validates :rule, presence: true
   validates :rule_reference, presence: true, uniqueness: { scope: :rule }
+
+  def self.find_unique(rule_references_rules)
+    arel_find(rule_references_rules) do |rule_references_rule|
+      arel_table[:rule_id].eq(rule_references_rule.rule_id).and(
+        arel_table[:rule_reference_id].eq(
+          rule_references_rule.rule_reference_id
+        )
+      )
+    end
+  end
 end