--- a/app/services/concerns/xccdf/test_result.rb
+++ b/app/services/concerns/xccdf/test_result.rb
@@ -4,7 +4,7 @@ module Xccdf
   # Methods related to saving TestResult from openscap_parser
   module TestResult
     def save_test_result
-      @test_result = ::TestResult.create(
+      @test_result = ::TestResult.create!(
         host: @host,
         profile: @host_profile,
         score: @op_test_result.score,