--- a/test/models/rule_reference_test.rb
+++ b/test/models/rule_reference_test.rb
@@ -29,6 +29,6 @@ class RuleReferenceTest < ActiveSupport::TestCase
   test 'finds an existing RuleReference from_oscap' do
     rule_reference = RuleReference.create!(OP_RULE_REFERENCE.to_h)
     assert_equal rule_reference.id,
-                 RuleReference.find_from_oscap([OP_RULE_REFERENCE]).first.id
+                 RuleReference.find_unique([OP_RULE_REFERENCE]).first.id
   end
 end