--- a/test/services/concerns/xccdf/rule_references_rules_test.rb
+++ b/test/services/concerns/xccdf/rule_references_rules_test.rb
@@ -19,7 +19,9 @@ class RuleReferencesRulesTest < ActiveSupport::TestCase
       save_rule_references_rules
     end
 
-    @rule_references_rules = nil # un-cache it from ||=
+    @new_rule_references_rules,
+      @existing_rule_references_rules,
+      @op_rule_references_rules = nil # un-cache it from ||=
     @op_rules = [
       OpenStruct.new(id: rules(:one).ref_id,
                      rule_references: [rule_references(:one)]),
@@ -30,6 +32,9 @@ class RuleReferencesRulesTest < ActiveSupport::TestCase
     assert_difference('RuleReferencesRule.count', 1) do
       save_rule_references_rules
     end
+
+    @new_rule_references_rules,
+      @existing_rule_references_rules = nil # un-cache it from ||=
 
     assert_difference('RuleReferencesRule.count', 0) do
       save_rule_references_rules