--- a/test/services/concerns/xccdf/rule_references_test.rb
+++ b/test/services/concerns/xccdf/rule_references_test.rb
@@ -17,7 +17,7 @@ class RuleReferencesTest < ActiveSupport::TestCase
       save_rule_references
     end
 
-    @rule_references, @new_rule_references = nil # un-cache it from ||=
+    @existing_rule_references, @new_rule_references = nil # un-cache it from ||=
     @op_rule_references = [OpenStruct.new(label: 'foo', href: ''),
                            OpenStruct.new(label: 'bar', href: '')]
     stubs(:new_rules).returns(