This change essentially boils down to removing N `selects` for N / 1000 (due to batch size). In order to accurately parse a report, we must determine if the records already exist in the DB. Before, we were using `find_or_initialize_by` for each RuleReference and RuleReferencesRule (37K records just for this table in the RHEL 7 0.1.46 SSG). Now we are executing a where query to determine existing records and only importing new records.

Importing RHEL 7 SSG, version 0.1.46, initial and subsequent imports. Initial import takes about half the time:

```
rake ssg:import DATASTREAM_FILE='./scap-security-guide-0.1.46/ssg-rhel7-ds.xml'

# Before

Importing ./scap-security-guide-0.1.46/ssg-rhel7-ds.xml at 2020-07-30 04:14:25 UTC
Finished importing ./scap-security-guide-0.1.46/ssg-rhel7-ds.xml in 86.885234702 seconds.

Importing ./scap-security-guide-0.1.46/ssg-rhel7-ds.xml at 2020-07-30 04:16:10 UTC
Finished importing ./scap-security-guide-0.1.46/ssg-rhel7-ds.xml in 7.394536394 seconds.

# After

Importing ./scap-security-guide-0.1.46/ssg-rhel7-ds.xml at 2020-07-30 04:10:54 UTC
Finished importing ./scap-security-guide-0.1.46/ssg-rhel7-ds.xml in 44.367159831 seconds.

Importing ./scap-security-guide-0.1.46/ssg-rhel7-ds.xml at 2020-07-30 04:12:20 UTC
Finished importing ./scap-security-guide-0.1.46/ssg-rhel7-ds.xml in 7.29955355 seconds.
```

There is more room for improvement during parsing. The same thing done here with RuleReferences and RuleReferencesRules, our biggest tables, can be done for Rules as well. 

Signed-off-by: Andrew Kofink <akofink@redhat.com>