@akofink 2020-07-30
Updated to use a `Struct` instead of an `OpenStruct`.